services/image-analysis/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
python-jose==3.3.0
requests==2.31.0
redis==5.0.1
minio==7.2.0
pydantic==2.6.1
python-dotenv==1.0.0
